requirements.txt
opencv-python==4.10.0.84
Pillow==11.0.0
torch==2.5.1
torchvision==0.20.1
matplotlib==3.10.0
python-Levenshtein==0.26.1
scikit-learn==1.6.0
seaborn==0.13.2
tqdm
requests==2.25.1
